requirements-dev.txt
# Development Dependencies for ShadowScan
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.1
pytest-xdist>=3.3.1
pytest-html>=3.2.0
requests-mock>=1.11.0
factory-boy>=3.3.0

# Code Quality & Linting
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.3.0
bandit[toml]>=1.7.5
safety>=2.3.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0
myst-parser>=2.0.0
sphinx-click>=5.0.0

# Build & Packaging
build>=0.10.0
twine>=4.0.0
setuptools>=68.0.0
wheel>=0.41.0

# Development Tools
ipython>=8.14.0
ipdb>=0.13.13
pdbpp>=0.10.3
watchdog>=3.0.0

# Profiling & Performance
memory-profiler>=0.61.0
py-spy>=0.3.14
line-profiler>=4.1.0

# Type Checking Extensions
types-requests>=2.31.0
types-PyYAML>=6.0.0
types-tabulate>=0.9.0
types-colorama>=0.4.15

# Development Utilities
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.5.0
click>=8.1.0

# Security Testing
semgrep>=1.45.0
pycryptodome>=3.18.0

# Performance Testing
locust>=2.16.0

# Container & Deployment Testing
docker>=6.1.0
docker-compose>=1.29.0

# Database Testing (if needed)
pytest-postgresql>=5.0.0
pytest-redis>=3.0.2

# Mock Services
responses>=0.23.0
httpretty>=1.1.4
freezegun>=1.2.2

# Code Coverage Reporting
coverage[toml]>=7.2.0
codecov>=2.1.13

# Static Analysis
pylint>=2.17.0
vulture>=2.9.1
mccabe>=0.7.0

# Documentation Testing
doctest>=3.11.0
pydocstyle>=6.3.0

# Git Hooks & Automation
detect-secrets>=1.4.0
gitpython>=3.1.32

# Development Environment
tox>=4.6.0
nox>=2023.4.22
virtualenv>=20.24.0

# Jupyter (for development notebooks)
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0

# Additional Development Dependencies
pipdeptree>=2.10.0
pip-tools>=7.1.0
bump2version>=1.0.1

# Performance Monitoring
psutil>=5.9.0

# Async Development
aiofiles>=23.1.0
aioresponses>=0.7.4

# Web Development Testing
selenium>=4.11.0
playwright>=1.36.0
httpx>=0.24.0

# Mock External APIs
vcrpy>=5.1.0
betamax>=0.8.1
